Fixed calculations values

+++ Propofol/CMCalculator.h
@@ -18,7 +18,7 @@
 
 -(void)logValues;
 
--(NSMutableDictionary *)newPatientWithAge:(int)age andWeight:(int)weight andHeight:(int)height andMale:(BOOL)male;
+-(NSMutableDictionary *)newPatientWithAge:(float)age andWeight:(float)weight andHeight:(float)height andMale:(BOOL)male;
 -(NSMutableDictionary *)giveDrugWithQuantity:(int)milligrams withState:(NSMutableDictionary *)state;
 -(NSMutableDictionary *)waitTime:(float)time withState:(NSMutableDictionary *)s;
